Created main class for testing purposes, created first draft of get All Links function
Finished getAllLinks function. Created a private function getBodyText as a helper to collect all text. getAllLinks uses this to initialize all hashmaps. Note that some hyperlinks don't go to a main article, but more of a "references"/"LEARN ABOUT THIS TOPIC in these articles"... article. I made sure the code accommodated for this by picking ALL topic paragraphs on the main body. Can fix if needed.

The hashmap called "bodyText" can be used for the idf/cos similarity calculations. Key is the article TITLE and value is the topic paragraph texts as one long string.